Module trait clean up
In particular, the test_relations functionality is superceded by the
version in FDModule that only checks generating relations.
